+++ simplyFact/database/migrations/2026_04_28_082127_create_training_expenses_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
                 ->constrained()
                 ->cascadeOnDelete();
             $table->integer('nb_days_of_training');
-            $table->float('reimbursed_price');
+            $table->unsignedInteger('reimbursed_price');
             $table->timestamps();
         });
     }
